Added Spotify Auth, Music brainz update
Spotify ios sdk used to sign in to spotify and saves the access token within the SpotifyAuthManager. About has a sign in/sign out button, playlist preview displays a "Sign in to generate playlist" when user is not signed in.

Music brainz service also updated to return artists with spotify Ids. makes it easily accessible to fetch music of those artists.